rename src to loftr so the package is installable with a reasonable name
