+++ configs/loftr/indoor/loftr_ds_dense.py
@@ -1,4 +1,4 @@
-from src.config.default import _CN as cfg
+from loftr.config.default import _CN as cfg
 
 cfg.LOFTR.MATCH_COARSE.MATCH_TYPE = 'dual_softmax'
