+++ configs/loftr/indoor/scannet/loftr_ds_eval.py
@@ -8,7 +8,7 @@ caused by tiny result fluctuation of few image pairs. This config set `BORDER_RM
 to be consistent with the results in our paper.
 """
 
-from src.config.default import _CN as cfg
+from loftr.config.default import _CN as cfg
 
 cfg.LOFTR.COARSE.TEMP_BUG_FIX = False
 cfg.LOFTR.MATCH_COARSE.MATCH_TYPE = 'dual_softmax'
