+++ configs/loftr/indoor/scannet/loftr_ds_eval_new.py
@@ -10,7 +10,7 @@ to be consistent with the results in our paper.
 Update: This config is for testing the re-trained model with the pos-enc bug fixed.
 """
 
-from src.config.default import _CN as cfg
+from loftr.config.default import _CN as cfg
 
 cfg.LOFTR.COARSE.TEMP_BUG_FIX = True
 cfg.LOFTR.MATCH_COARSE.MATCH_TYPE = 'dual_softmax'
