+++ loftr/lightning/lightning_loftr.py
@@ -9,19 +9,19 @@ import numpy as np
 import pytorch_lightning as pl
 from matplotlib import pyplot as plt
 
-from src.loftr import LoFTR
-from src.loftr.utils.supervision import compute_supervision_coarse, compute_supervision_fine
-from src.losses.loftr_loss import LoFTRLoss
-from src.optimizers import build_optimizer, build_scheduler
-from src.utils.metrics import (
+from loftr.loftr import LoFTR
+from loftr.loftr.utils.supervision import compute_supervision_coarse, compute_supervision_fine
+from loftr.losses.loftr_loss import LoFTRLoss
+from loftr.optimizers import build_optimizer, build_scheduler
+from loftr.utils.metrics import (
     compute_symmetrical_epipolar_errors,
     compute_pose_errors,
     aggregate_metrics
 )
-from src.utils.plotting import make_matching_figures
-from src.utils.comm import gather, all_gather
-from src.utils.misc import lower_config, flattenList
-from src.utils.profiler import PassThroughProfiler
+from loftr.utils.plotting import make_matching_figures
+from loftr.utils.comm import gather, all_gather
+from loftr.utils.misc import lower_config, flattenList
+from loftr.utils.profiler import PassThroughProfiler
 
 
 class PL_LoFTR(pl.LightningModule):
