+++ requirements.txt
@@ -1,5 +1,6 @@
 opencv_python==4.4.0.46
-albumentations==0.5.1 --no-binary=imgaug,albumentations
+albumentations==0.5.1
+imgaug==0.4.0 
 # ray>=1.0.1
 einops==0.3.0
 kornia==0.4.1
