+++ test.py
@@ -3,11 +3,11 @@ import argparse
 import pprint
 from loguru import logger as loguru_logger
 
-from src.config.default import get_cfg_defaults
-from src.utils.profiler import build_profiler
+from loftr.config.default import get_cfg_defaults
+from loftr.utils.profiler import build_profiler
 
-from src.lightning.data import MultiSceneDataModule
-from src.lightning.lightning_loftr import PL_LoFTR
+from loftr.lightning.data import MultiSceneDataModule
+from loftr.lightning.lightning_loftr import PL_LoFTR
 
 
 def parse_args():
